fix: use DropShadowOnly only for background.
The thing is DropShadowOnly will ignore the content, that's fine for background, since we don't to draw the rectangle along the shadow. We only want to draw the shadow in the rectangle (use an invisible color for the paint does not work).

However, if it's a text node, we don't want to ignore the text, and we want to draw it.

The main difference is because to draw the text, we use the same paint, but to draw the background we use different paints.

Keep in mind we decided to use different paint objects for the background since we want to be able to draw the shadow either with bg color or bg image.

Finally, keep in mind that using the same paint to draw the bg image and the shadows were causing some issues, so that's not a valid alternative.