fix(bounds): use float bounds + AntiAlias=True arg in skia.Paint, causes lines of pixels using colors with alpha channel:
Suppose bounds (0, 0, 10.5, 10.5) + AntiAlias=True, with solid color "red". It will cause that the X-column number 11 will be painted with 0.5 (127) of alpha cannel because of the AntiAlias. The same for the Y-row number 11.

To fix that, we decided to use int bounds: if one element "touches" a single part of a pixel in a column/row, then that whole column/row is inside the bounds.